appbuilder/requirements.txt
# AppBuilder Dependencies
google-genai>=0.4.0
groq>=0.13.0
PyGithub>=2.1.1
discord.py>=2.3.2
python-dotenv>=1.0.0
typing-extensions>=4.10.0
requests>=2.31.0
